partial rework of system as obj
