+++ lib/benchpark/spec.py
@@ -6,10 +6,10 @@ import benchpark.paths
 import benchpark.repo
 import benchpark.runtime
 
-import llnl.util.lang
-
 bootstrapper = benchpark.runtime.RuntimeResources(benchpark.paths.benchpark_home)
 bootstrapper.bootstrap()
+
+import llnl.util.lang
 
 repo_path = benchpark.repo.paths[benchpark.repo.ObjectTypes.experiments]
 
@@ -196,6 +196,20 @@ class ExperimentSpec(Spec):
 
     def concretize(self):
         return ConcreteExperimentSpec(self)
+
+
+def autospec(function):
+    """Decorator that automatically converts the first argument of a
+    function to a Spec.
+    """
+
+    @functools.wraps(function)
+    def converter(self, spec_like, *args, **kwargs):
+        if not isinstance(spec_like, ExperimentSpec):
+            spec_like = ExperimentSpec(spec_like)
+        return function(self, spec_like, *args, **kwargs)
+
+    return converter
 
 
 class ConcreteSpec(Spec):
