+++ lib/benchpark/systems/aws/system.py
@@ -3,7 +3,7 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-from .base import System
+from benchpark.system import System
 
 # Taken from https://aws.amazon.com/ec2/instance-types/
 # With boto3, we could determine this dynamically vs. storing a static table
