+++ lib/benchpark/systems/tioga/externals/rocm/00-version-543-packages.yaml
